fix: FE와의 연결 및 버그 수정

+++ src/main/java/boot/kakaotech/communitybe/post/dto/PostListDto.java
@@ -17,9 +17,9 @@ public class PostListDto {
 
     private String title;
 
-    private int likeCount;
+    private long likeCount;
 
-    private int commentCount;
+    private long commentCount;
 
     private int viewCount;
